multiphy: Allow clients to snap to physical screen edges
Comparisons are only tested against the edges of the physical screen
that the client currently resides in.

Ie, if (as xinerama/xrandr allows), there were a gap between two
horizontally-adjacent screens (L, R), and a client (C) is dragged
from L to R, the right-hand edge of C will only snap against the
edge A, not B.[1]

       L            R
   .--------.  .--------.
   |        A  B        |
   |  CCC   A  B        |
   |  CCC   A  B        |
   |        A  B        |
   `--------'  `--------'

[1], the technical exception to this is if the gap between L and R
is sufficiently large so as to allow the notion of which physical
screen the client resides to change[2] before the right-hand edge
of C reaches B.

[2] Currently, when the centre of the client window is not on a
physical screen, membership is taken as the screen with the closest
centre.