multiphy: Allow vdesk per physical screen
An (X)screen has a pool of vdesks that can be mapped to any physical
screen.  All windows will be positioned in the same relative place
to the new phsyical screen's origin.

NB, not possible to display the same vdesk on multiple physical screens
at the same time. This reduces the number of effective vdesks, since each
physical display is always displaying a vdesk.