fix: handle failure of disabling auto-repeat when displaying infobanner
Disabling keyboard auto-repeat may not have any effect with some X
servers.  The change in auto-repeat mode may take effect after all
keys have been released; this is too late for the infobanner routine.

When the X server generates the fake KeyRelease and KeyPress events
for auto-repeat, the event time will be identical for each pair.
This may be used to detect and discard KeyPress events that are
part of the auto-repeat.