fix: race condition when unmapping windows causing unmaged windows.
When unmapping multiple windows, it was possible for the multiple
clients to be unmapped at once, the batch of events denoting unmapping
can include enter events.  Evilwm handles each event in turn without
any look-ahead, so blindly calls XSetInputFocus on a window that has
been unmapped (forwich the enter event is out-of-date).

This bug was spotted with a dock and a window a tool on the dock created.
It should be possible with vdesk switching too.

The trivial fix is to ignore Xerrors for such a trivial error, rather
than disowning the client.